updated strategy
